added a cluster file to keep track of cluster info. Moved the bond network to graph code here, but have not modified anything yet.
