+++ src/analyzeStructures_refactor.py
@@ -41,6 +41,7 @@ import os
 
 from body import body
 from body import neighborgrid as ng
+from body import cluster as cluster
 
 #this is a hack for now. make this more general later. WARNING: increase this size to be 
 #larger than the largest structure you see in your simulations
@@ -366,112 +367,6 @@ class SimInfo:
 
 
 ####################################################################
-################# Bond Network -> Graph & Clustering ###############
-####################################################################
-
-def get_groups(bond_dict):
-    #construct arrays containing groups of all bonded particles
-
-    #get 'state' info
-    states = list(bond_dict.keys())
-    total_states = len(states)
-
-    #if no bonds, return 0
-    if total_states == 0:
-        return [], bond_dict
-
-    #convert the indices in the dict to 1:total_states
-    new_bond_dict = dict()
-    for i in range(total_states):
-
-        #get the list from the original bond_dict
-        entries = bond_dict[states[i]]
-
-        #convert the entries to new indices
-        new_entries = []
-        for entry in entries:
-            new_entries.append(np.where(states == entry)[0][0])
-
-        #add new entries to new dict
-        new_bond_dict[i] = new_entries
-
-    #replace old dict
-    bond_dict = new_bond_dict
-
-    #init an array to store mappings to groups
-    to_group = -1 * np.ones(total_states, dtype=int)
-    to_group[0] = 0
-
-    #init a list to store the groups
-    G = [[0]]
-
-    #init a queue and dump for searching bonds
-    queue    = [0]
-    searched = []
-
-    #loop over the queue until it is empty or all bonds have been assigned
-    while len(queue) > 0:
-
-        #pop the front of the queue and get its group number
-        particle = queue.pop(0)
-        group_num = to_group[particle]
-
-        #get the adjacency from bond_dict for this particle
-        adj = bond_dict[particle]
-
-        #for each adjacent particle, add it to the group, set the map, add to queue
-        for member in adj:
-
-            #if the member is not in the group yet, add it and set the mapping
-            if member not in G[group_num]:
-                G[group_num].append(member)
-                to_group[member] = group_num
-
-            #if the member has not been searched and is not in queue, add
-            if member not in searched and member not in queue:
-                queue.append(member)
-
-        #append the particle to the searched array
-        searched.append(particle)
-
-        #check if the queue is empty but there are undetermined states
-        if len(queue) == 0 and len(searched) != total_states:
-
-            #determine the first unassigned state
-            unassigned = np.where(to_group == -1)[0][0]
-
-            #create a new group for it, assign it, append to queue
-            G.append([unassigned])
-            to_group[unassigned] = len(G)-1
-            queue.append(unassigned)
-
-    #return the list of grouped particles
-    return G, bond_dict
-
-
-def get_group_sizes(G):
-
-    #init an array of zeros up to MAX_SIZE
-    size_counts = np.zeros(MAX_SIZE, dtype=int)
-
-    #loop over groups and increment the corresponding size index
-    for group in G:
-
-        L = len(group)
-        size_counts[L] += 1
-
-    #determine the largest group
-    non_zero_counts = np.where(size_counts > 0)[0]
-    if len(non_zero_counts) > 0:
-        largest_group_size = np.max(non_zero_counts)
-    else:
-        largest_group_size = 0
-
-    #return size counts
-    return size_counts, largest_group_size
-
-
-####################################################################
 ################# Main Drivers #####################################
 ####################################################################
 
@@ -489,15 +384,19 @@ def analyze_structures(snap, sim, radius = None, center = None):
 
     #determine the bond network using the list of bodies
     body.get_bonded_bodies(bodies, sim, bond_dict)
+
+    #debug lines
     # print(bond_dict)
-    bonds = 0
-    for key in bond_dict.keys():
-        bonds += len(bond_dict[key])
-    print("Bonds ", bonds/2)
-    return bonds, bonds
+    # bonds = 0
+    # for key in bond_dict.keys():
+    #     bonds += len(bond_dict[key])
+    # print("Bonds ", bonds/2)
+    # return bonds, bonds
 
     #determine groups of bonded structures
-    G, bond_dict = get_groups(bond_dict)
+    G, bond_dict = cluster.get_groups(bond_dict)
+    print(G)
+    sys.exit()
 
     #count the sizes of each group
     size_counts, largest_group_size = get_group_sizes(G)
